requirements.txt
meshtastic==2.2.20
pillow
requests
pyserial
pypubsub
